+++ Gemfile
@@ -1,4 +1,4 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in jahuty-snippets.gemspec
+# Specify your gem's dependencies in jahuty.gemspec
 gemspec
