+++ lib/jahuty.rb
@@ -1,0 +1,12 @@
+require "jahuty/version"
+
+require "jahuty/data/problem"
+require "jahuty/data/snippet"
+
+require "jahuty/exception/not_ok"
+
+require "jahuty/service/connect"
+require "jahuty/service/get"
+
+module Jahuty
+end
