+++ lib/jahuty/service/get.rb
@@ -1,0 +1,28 @@
+require "json"
+
+module Jahuty
+  module Service
+    class Get
+      @connection
+
+      def initialize(connection)
+        @connection = connection
+
+      end
+
+      def call(id)
+        response = @connection.get(id.to_s)
+
+        payload = JSON.parse(response.body, symbolize_names: true)
+
+        if response.status != 200
+          raise ::Jahuty::Exception::NotOk.new(
+            ::Jahuty::Data::Problem.from(payload)
+          )
+        end
+
+        return ::Jahuty::Data::Snippet.from(payload)
+      end
+    end
+  end
+end
