--- lib/jahuty/snippet/version.rb
@@ -1,5 +1,0 @@
-module Jahuty
-  module Snippet
-    VERSION = "0.1.0"
-  end
-end
