+++ lib/jahuty/version.rb
@@ -1,0 +1,3 @@
+module Jahuty
+  VERSION = "1.0.0"
+end
