+++ spec/jahuty/exception/not_ok_spec.rb
@@ -1,0 +1,18 @@
+module Jahuty
+  module Exception
+    RSpec.describe NotOk do
+      describe "#message" do
+        it "returns message with problem's status, title, and detail" do
+          status = 1
+          title  = "foo"
+          detail = "bar"
+
+          problem = ::Jahuty::Data::Problem.new(status, title, detail)
+          subject = NotOk.new(problem)
+
+          expect(subject.message).to include(status.to_s, title, detail)
+        end
+      end
+    end
+  end
+end
