+++ spec/jahuty_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe Jahuty do
+  it "has a version number" do
+    expect(Jahuty::VERSION).not_to be nil
+  end
+end
